upd: dataLoatBai, dataVideos MainContext
